refactor: move crypto utility state to a separete compostive
