+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/Productizer/Operations/GetUser.cs
@@ -48,7 +48,7 @@ public static class GetUser
 
         public async Task<User> Handle(Query request, CancellationToken cancellationToken)
         {
-            _usersDbContext.Cryptor.StartQuery("Person", request.EncryptionKey);
+            _usersDbContext.Cryptor.State.StartQuery("Person", request.EncryptionKey);
             var dbUser = await _usersDbContext.Persons
                 .Include(p => p.Occupations)
                 .Include(p => p.WorkPreferences)
