+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/Productizer/Operations/UpdateUser.cs
@@ -119,7 +119,7 @@ public static class UpdateUser
 
         public async Task<User> Handle(Command request, CancellationToken cancellationToken)
         {
-            _usersDbContext.Cryptor.StartQuery("Person", request.EnryptionKey);
+            _usersDbContext.Cryptor.State.StartQuery("Person", request.EnryptionKey);
             var dbUser = await _usersDbContext.Persons
                 .Include(p => p.AdditionalInformation).ThenInclude(ai => ai!.Address)
                 .SingleAsync(o => o.Id == request.UserId, cancellationToken);
