+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Activities/User/Operations/UpdateUser.cs
@@ -149,7 +149,7 @@ public static class UpdateUser
 
         public async Task<User> Handle(Command request, CancellationToken cancellationToken)
         {
-            _usersDbContext.Cryptor.StartQuery("Person", request.EncryptionKey);
+            _usersDbContext.Cryptor.State.StartQuery("Person", request.EncryptionKey);
             var dbUser = await _usersDbContext.Persons
                 .Include(u => u.WorkPreferences)
                 .Include(u => u.Occupations)
